Rename clustering_plots to clustering.
