+++ docs/modules/exploratory/visualization.rst
@@ -10,4 +10,4 @@ the functional data, that highlight several important aspects of it.
    
    visualization/boxplot
    visualization/magnitude_shape_plot
-   visualization/clustering_plots
+   visualization/clustering
