+++ examples/plot_clustering.py
@@ -15,7 +15,7 @@ calculate the results plotted.
 import matplotlib.pyplot as plt
 import numpy as np
 from skfda import datasets
-from skfda.exploratory.visualization.clustering_plots import (
+from skfda.exploratory.visualization.clustering import (
     plot_clusters, plot_cluster_lines, plot_cluster_bars)
 from skfda.ml.clustering.base_kmeans import KMeans, FuzzyKMeans
