+++ skfda/exploratory/visualization/__init__.py
@@ -1,3 +1,3 @@
+from . import clustering
 from ._boxplot import Boxplot, SurfaceBoxplot
 from ._magnitude_shape_plot import MagnitudeShapePlot
-from . import clustering_plots
